+++ Exodus/src/utopia/exodus/rest/resource/user/me/DeletionsForMyOrganizationsNode.scala
@@ -1,4 +1,4 @@
-package utopia.exodus.rest.resource.user
+package utopia.exodus.rest.resource.user.me
 
 import utopia.access.http.Status.NotImplemented
 import utopia.exodus.rest.util.AuthorizedContext
@@ -7,10 +7,10 @@ import utopia.nexus.rest.ResourceWithChildren
 import utopia.nexus.result.Result
 
 /**
-  * Used for accessing deletions concerning the organizations the current user is a member of
-  * @author Mikko Hilpinen
-  * @since 16.5.2020, v1
-  */
+ * Used for accessing deletions concerning the organizations the current user is a member of
+ * @author Mikko Hilpinen
+ * @since 16.5.2020, v1
+ */
 object DeletionsForMyOrganizationsNode extends ResourceWithChildren[AuthorizedContext]
 {
 	override def children = Vector(PendingDeletionsForMyOrganizationsNode)