+++ Exodus/src/utopia/exodus/rest/resource/user/me/MeNode.scala
@@ -1,4 +1,4 @@
-package utopia.exodus.rest.resource.user
+package utopia.exodus.rest.resource.user.me
 
 import utopia.access.http.Method.Get
 import utopia.access.http.Status.Unauthorized
@@ -10,25 +10,24 @@ import utopia.nexus.result.Result
 import utopia.vault.database.Connection
 
 /**
-  * This rest-resource represents the logged user
-  * @author Mikko Hilpinen
-  * @since 6.5.2020, v1
-  */
+ * This rest-resource represents the logged user
+ * @author Mikko Hilpinen
+ * @since 6.5.2020, v1
+ */
 object MeNode extends ExtendableSessionResource
 {
 	override val name = "me"
 	
 	private val defaultGet = SessionUseCaseImplementation.default(Get) { (session, connection, _, _) =>
-			implicit val c: Connection = connection
-			// Reads user data and adds linked data
-			DbUser(session.userId).withLinks match
-			{
-				case Some(user) => Result.Success(user.toModel)
-				case None =>
-					// Log.warning(s"User id ${session.userId} was authorized but couldn't be found from the database")
-					Result.Failure(Unauthorized, "User no longer exists")
-			}
+		implicit val c: Connection = connection
+		// Reads user data and adds linked data
+		DbUser(session.userId).withLinks match {
+			case Some(user) => Result.Success(user.toModel)
+			case None =>
+				// Log.warning(s"User id ${session.userId} was authorized but couldn't be found from the database")
+				Result.Failure(Unauthorized, "User no longer exists")
 		}
+	}
 	
 	override protected val defaultUseCaseImplementations = Vector(defaultGet)
 	override protected val defaultFollowImplementations =