+++ Exodus/src/utopia/exodus/rest/resource/user/me/MyLanguagesNode.scala
@@ -1,4 +1,4 @@
-package utopia.exodus.rest.resource.user
+package utopia.exodus.rest.resource.user.me
 
 import utopia.access.http.Method.{Delete, Get, Post, Put}
 import utopia.access.http.Status.{BadRequest, Forbidden}
@@ -19,10 +19,10 @@ import utopia.vault.database.Connection
 import scala.util.{Failure, Success}
 
 /**
-  * Used for interacting with the languages known to the authorized user
-  * @author Mikko Hilpinen
-  * @since 16.5.2020, v1
-  */
+ * Used for interacting with the languages known to the authorized user
+ * @author Mikko Hilpinen
+ * @since 16.5.2020, v1
+ */
 object MyLanguagesNode extends Resource[AuthorizedContext]
 {
 	// IMPLEMENTED	-------------------------
@@ -39,8 +39,7 @@ object MyLanguagesNode extends Resource[AuthorizedContext]
 			val user = DbUser(session.userId)
 			
 			// GET simply returns existing user language links (with appropriate descriptions)
-			if (method == Get)
-			{
+			if (method == Get) {
 				// Reads language descriptions from the database, uses languages in the accept-language header
 				val languages = user.languages.withDescriptionsInLanguages(context.languageIdListFor(session.userId))
 					.sortBy { _.familiarity.orderIndex }
@@ -53,35 +52,30 @@ object MyLanguagesNode extends Resource[AuthorizedContext]
 				})
 			}
 			// DELETE removes known languages
-			else if (method == Delete)
-			{
+			else if (method == Delete) {
 				context.handleArrayPost { values =>
 					val existingLanguageIds = user.languageIds.toSet
 					val languageIdsToDelete = values.flatMap { _.int }.toSet
 					// Makes sure at leas one language remains
 					if (existingLanguageIds.forall(languageIdsToDelete.contains))
 						Result.Failure(Forbidden, "You must keep at least one language")
-					else
-					{
+					else {
 						user.languages.remove(languageIdsToDelete)
 						// Returns a list of remaining language ids
 						Result.Success((existingLanguageIds -- languageIdsToDelete).toVector)
 					}
 				}
 			}
-			else
-			{
+			else {
 				context.handleModelArrayPost(NewLanguageProficiency) { proficiencies =>
 					// Validates the proposed languages first
-					DbLanguage.validateProposedProficiencies(proficiencies) match
-					{
+					DbLanguage.validateProposedProficiencies(proficiencies) match {
 						case Success(proficiencies) =>
 							// on POST, adds new language proficiencies (may overwrite existing)
 							// on PUT, replaces languages
 							if (proficiencies.isEmpty && method == Put)
 								Result.Failure(Forbidden, "You must keep at least one language")
-							else
-							{
+							else {
 								val existingLanguages = user.languages.withFamiliarityLevels.toMap
 								
 								// Groups the changes
@@ -89,21 +83,25 @@ object MyLanguagesNode extends Resource[AuthorizedContext]
 								
 								val changesInExisting = existingLanguages.flatMap { case (languageId, familiarity) =>
 									changesMap.get(languageId).map { newFamiliarity =>
-										languageId -> (familiarity -> newFamiliarity) } }
+										languageId -> (familiarity -> newFamiliarity)
+									}
+								}
 								val modifications = changesInExisting
 									.filterNot { case (_, change) => change._1.id == change._2 }
 								val duplicateLanguageIds = changesInExisting.filter { case (_, change) =>
-									change._1.id == change._2 }.keySet
+									change._1.id == change._2
+								}.keySet
 								val newLanguages = changesMap.filterNot { case (languageId, _) =>
-									existingLanguages.contains(languageId) }
+									existingLanguages.contains(languageId)
+								}
 								
 								// Removes some languages (those not listed in PUT and those being modified)
-								val languageIdsToRemove =
-								{
+								val languageIdsToRemove = {
 									val base = modifications.keySet
 									if (method == Put)
 										base ++ existingLanguages.filterNot { case (languageId, _) =>
-											changesMap.contains(languageId) }.keySet
+											changesMap.contains(languageId)
+										}.keySet
 									else
 										base
 								}
@@ -111,24 +109,31 @@ object MyLanguagesNode extends Resource[AuthorizedContext]
 								
 								// Adds new language links
 								val inserted = (modifications.map { case (languageId, change) =>
-									languageId -> change._2 } ++ newLanguages).map { case (languageId, familiarity) =>
-									user.languages.insert(languageId, familiarity) }
+									languageId -> change._2
+								} ++ newLanguages).map { case (languageId, familiarity) =>
+									user.languages.insert(languageId, familiarity)
+								}
 								
 								// Returns the new languages list
 								val duplicates = existingLanguages.filter { case (languageId, _) =>
-									duplicateLanguageIds.contains(languageId) }
+									duplicateLanguageIds.contains(languageId)
+								}
 								val duplicateModels = duplicates.map { case (languageId, familiarity) =>
-									Model(Vector("language_id" -> languageId, "familiarity_id" -> familiarity.id)) }.toVector
-								val insertedModels = inserted.map { l => Model(Vector("language_id" -> l.languageId,
-									"familiarity_id" -> l.familiarityId)) }.toVector
+									Model(Vector("language_id" -> languageId, "familiarity_id" -> familiarity.id))
+								}.toVector
+								val insertedModels = inserted.map { l =>
+									Model(Vector("language_id" -> l.languageId,
+										"familiarity_id" -> l.familiarityId))
+								}.toVector
 								
 								if (method == Put)
 									Result.Success(duplicateModels ++ insertedModels)
-								else
-								{
+								else {
 									val previousModels = existingLanguages.filterNot { case (languageId, _) =>
-										changesMap.contains(languageId) }.map { case (languageId, familiarity) =>
-										Model(Vector("language_id" -> languageId, "familiarity_id" -> familiarity.id)) }.toVector
+										changesMap.contains(languageId)
+									}.map { case (languageId, familiarity) =>
+										Model(Vector("language_id" -> languageId, "familiarity_id" -> familiarity.id))
+									}.toVector
 									Result.Success(previousModels ++ duplicateModels ++ insertedModels)
 								}
 							}