+++ Exodus/src/utopia/exodus/rest/resource/user/me/MyOrganizationsNode.scala
@@ -1,4 +1,4 @@
-package utopia.exodus.rest.resource.user
+package utopia.exodus.rest.resource.user.me
 
 import utopia.access.http.Method.Get
 import utopia.citadel.database.access.many.description.DbDescriptionRoles
@@ -12,10 +12,10 @@ import utopia.nexus.result.Result
 import utopia.vault.database.Connection
 
 /**
-  * This rest node returns a descriptive list of all the organizations the authorized user belongs to
-  * @author Mikko Hilpinen
-  * @since 6.5.2020, v1
-  */
+ * This rest node returns a descriptive list of all the organizations the authorized user belongs to
+ * @author Mikko Hilpinen
+ * @since 6.5.2020, v1
+ */
 object MyOrganizationsNode extends ResourceWithChildren[AuthorizedContext]
 {
 	// IMPLEMENTED	------------------------
@@ -33,12 +33,10 @@ object MyOrganizationsNode extends ResourceWithChildren[AuthorizedContext]
 			// Reads organizations data and returns it as an array
 			// Also supports the If-Modified-Since / Not Modified use case
 			DbUser(session.userId).memberships.myOrganizations(
-				context.languageIdListFor(session.userId), context.request.headers.ifModifiedSince) match
-			{
+				context.languageIdListFor(session.userId), context.request.headers.ifModifiedSince) match {
 				case Some(organizations) =>
 					// May use simple model style
-					Result.Success(session.modelStyle match
-					{
+					Result.Success(session.modelStyle match {
 						case Full => organizations.map { _.toModel }
 						case Simple =>
 							val descriptionRoles = DbDescriptionRoles.all