+++ Exodus/src/utopia/exodus/rest/resource/user/me/PendingDeletionsForMyOrganizationsNode.scala
@@ -1,4 +1,4 @@
-package utopia.exodus.rest.resource.user
+package utopia.exodus.rest.resource.user.me
 
 import utopia.access.http.Method.Get
 import utopia.citadel.database.access.many.organization.DbOrganizations
@@ -12,10 +12,10 @@ import utopia.nexus.result.Result
 import utopia.vault.database.Connection
 
 /**
-  * Used for accessing pending deletions in the current user's organizations
-  * @author Mikko Hilpinen
-  * @since 16.5.2020, v1
-  */
+ * Used for accessing pending deletions in the current user's organizations
+ * @author Mikko Hilpinen
+ * @since 16.5.2020, v1
+ */
 object PendingDeletionsForMyOrganizationsNode extends Resource[AuthorizedContext]
 {
 	override val name = "pending"